authoscope 0.8.0
This formula was renamed from `badtouch`.
